requirements.txt
snowflake-connector-python>=3.0.0
pandas>=1.0.0
python-dotenv>=0.19.0
dune-client>=1.3.0
gunicorn>=20.1.0 